app/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.50.0
frozendict==2.4.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
html5lib==1.1
ibm-cos-sdk==2.11.0
ibm-cos-sdk-core==2.11.0
ibm-cos-sdk-s3transfer==2.11.0
ibm-watson-machine-learning==1.0.333
ibm_db==3.2.3
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lomond==0.3.3
lxml==5.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
multitasking==0.0.11
numpy==1.26.4
packaging==23.2
pandas==1.5.3
peewee==3.17.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
streamlit==1.32.2
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.29.0
watchdog==4.0.0
webencodings==0.5.1
Werkzeug==3.0.1
yfinance==0.2.37
zipp==3.18.1
